Add UrlReservation model including validation.
This code is mostly copied across directly from url-arbiter. The model has been
renamed from Reservation for clarity, since this app deals with more than just
URLs.

In addition, checking that the path can be claimed by that publishing app has
been moved from a method into a validator, so it can be caught as a validation
error in the commands.