GetLinkSet: return empty links for non-existing linksets
`GetLinkSet` is currently returning 404 when the there is no `LinkSet`
record with the specified `content_id`. This happens even when there is
a content item present with that `content_id`.

Because LinkSet records are only created when the first links are
created, this will happen regularly.

This response is not ideal because clients may assume that asking for
the links of an existing content item without links will return a links
hash.

This commit makes the API respond consistently to
`/v2/links/:content_id` requests.