Shoehorn message queue into the content store worker
This is a proof of concept to show the performance improvements we can
achieve by having the message queue and the content store done in the
same asynchronous operation.

The combining of them within this is pretty ugly, it has increased the
amount of options going into worker and it is now performing operations
outside of it's domain model.

It would be prudent to instead have a DownstreamWorker (or similarly
named) that is designed for the operations we want to perform
asynchronously and to consider what options we can pass in to make it
easy to understand.

Benchmarks from `$ ruby bench/publish.rb`

**This Branch**
Creating drafts...
Publishing...
....................................................................................................
  2.460000   0.320000   2.780000 (  3.559283)
2301 SQL queries
Creating new drafts...
Superseding...
....................................................................................................
  2.760000   0.440000   3.200000 (  4.087826)

**Master Branch**

Creating drafts...
Publishing...
....................................................................................................
  3.520000   0.640000   4.160000 (  5.840053)
3203 SQL queries
Creating new drafts...
Superseding...
....................................................................................................
  3.730000   0.550000   4.280000 (  5.927691)
3700 SQL queries