Preload the web content items using a recursive query during link expansion
This should reduce the number of queries done for content items a large
"transitive closure", at the cost of increasing the number of queries for
content items with a small "transitive closure".

At the moment, I have only done testing with bench/publish.rb, which is slower
and takes more queries with this change, which is expected as the content items
have no links.

At the moment, my preference is to not pursue this further until there is more
value to be gained.