Table lock to resolve race conditions
We hit a race condition when two versions of a content item are
submitted to the Publishing API simultaneously and there isn't any
historical versions existing.

This adds a lock when there is not a Content Item existing in a draft or
published form to prevent any simultaneous requests from completing
concurrently.