Event S3 Exporter class
Given date inputs this class will access all Events that have a nil
payload and are with the date ranges.

These events will be imported into a gzipped CSV file and this file will
be uploaded to S3 with the timestamp as a key. If a file already exists
an exception will be thrown.

If there are no items to export an empty file will be uploaded.